moved low-level logic of adding WebSocket listeners from the Room#watch to Client#watch
